+++ src/Audit/AuditOSSIndex.ts
@@ -13,8 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { OssIndexServerResult, Vulnerability } from '../Types/OssIndexServerResult';
-import chalk from 'chalk';
+import { OssIndexServerResult } from '../Types/OssIndexServerResult';
 import { Formatter, getNumberOfVulnerablePackagesFromResults } from './Formatters/Formatter';
 import { JsonFormatter } from './Formatters/JsonFormatter';
 import { TextFormatter } from './Formatters/TextFormatter';
@@ -29,110 +28,21 @@ export class AuditOSSIndex {
     } else if (xml) {
       this.formatter = new XmlFormatter();
     } else {
-      this.formatter = new TextFormatter();
+      this.formatter = new TextFormatter(quiet);
     }
   }
 
   public auditResults(results: Array<OssIndexServerResult>): boolean {
-    if (this.json) {
-      this.formatter.printAuditResults(results);
-      return getNumberOfVulnerablePackagesFromResults(results) > 0;
-    }
-    if (this.xml) {
-      this.formatter.printAuditResults(results);
-      return getNumberOfVulnerablePackagesFromResults(results) > 0;
+    if (this.quiet) {
+      results = results.filter((x) => {
+        return x.vulnerabilities && x.vulnerabilities?.length > 0;
+      });
     }
 
-    const total = results.length;
-    results = results.sort((a, b) => {
-      return a.coordinates < b.coordinates ? -1 : 1;
-    });
+    this.formatter.printAuditResults(results);
 
-    console.log();
-    console.group();
-    this.printLine('Sonabot here, beep boop beep boop, here are your Sonatype OSS Index results:');
-    this.suggestIncludeDevDeps(total);
-    console.groupEnd();
-    console.log();
-
-    this.printLine('-'.repeat(process.stdout.columns));
-
-    let isVulnerable = false;
-
-    results.forEach((x: OssIndexServerResult, i: number) => {
-      if (x.vulnerabilities && x.vulnerabilities.length > 0) {
-        isVulnerable = true;
-        this.printVulnerability(i, total, x);
-      } else {
-        this.printLine(chalk.keyword('green')(`[${i + 1}/${total}] - ${x.toAuditLog()}`));
-      }
-    });
-
-    this.printLine('-'.repeat(process.stdout.columns));
+    let isVulnerable = getNumberOfVulnerablePackagesFromResults(results) > 0;
 
     return isVulnerable;
-  }
-
-  private getColorFromMaxScore(maxScore: number, defaultColor = 'chartreuse'): string {
-    if (maxScore > 8) {
-      defaultColor = 'red';
-    } else if (maxScore > 6) {
-      defaultColor = 'orange';
-    } else if (maxScore > 4) {
-      defaultColor = 'yellow';
-    }
-    return defaultColor;
-  }
-
-  private printVulnerability(i: number, total: number, result: OssIndexServerResult): void {
-    if (!result.vulnerabilities) {
-      return;
-    }
-    const maxScore: number = Math.max(
-      ...result.vulnerabilities.map((x: Vulnerability) => {
-        return +x.cvssScore;
-      }),
-    );
-    const printVuln = (x: Array<Vulnerability>): void => {
-      x.forEach((y: Vulnerability) => {
-        const color: string = this.getColorFromMaxScore(+y.cvssScore);
-        console.group();
-        console.log(chalk.keyword(color)(`Vulnerability Title: `), `${y.title}`);
-        console.log(chalk.keyword(color)(`ID: `), `${y.id}`);
-        console.log(chalk.keyword(color)(`Description: `), `${y.description}`);
-        console.log(chalk.keyword(color)(`CVSS Score: `), `${y.cvssScore}`);
-        console.log(chalk.keyword(color)(`CVSS Vector: `), `${y.cvssVector}`);
-        console.log(chalk.keyword(color)(`CVE: `), `${y.cve}`);
-        console.log(chalk.keyword(color)(`Reference: `), `${y.reference}`);
-        console.log();
-        console.groupEnd();
-      });
-    };
-
-    console.log(
-      chalk.keyword(this.getColorFromMaxScore(maxScore)).bold(`[${i + 1}/${total}] - ${result.toAuditLog()}`),
-    );
-    console.log();
-    result.vulnerabilities &&
-      printVuln(
-        result.vulnerabilities.sort((x, y) => {
-          return +y.cvssScore - +x.cvssScore;
-        }),
-      );
-  }
-
-  private printLine(line: any): void {
-    if (!this.quiet) {
-      console.log(line);
-    }
-  }
-
-  private suggestIncludeDevDeps(total: number) {
-    this.printLine(`Total dependencies audited: ${total}`);
-    if (total == 0) {
-      this.printLine(
-        `We noticed you had 0 dependencies, we exclude devDependencies by default, try running with --dev if you want to include those as well`,
-      );
-    }
   }
 }
